modifications for classification, segmentation metrics
