+++ volkscv/metrics/__init__.py
@@ -1,0 +1,4 @@
+from .base import BaseMetric
+from .confusion_matrix import BaseConfusionMatrix
+
+__all__ = ['BaseMetric', 'BaseConfusionMatrix']
