+++ volkscv/metrics/classification/__init__.py
@@ -1,6 +1,5 @@
-from .base import BaseMetric
 from .accuracy import TopKAccuracy
-from .confusion_matrix import ConfusionMatrix, CMAccuracy, CMPrecisionRecall
+from .cfx_based_metrics import ConfusionMatrix, CMAccuracy, CMPrecisionRecall
 from .fbeta_score import Fbetascore
 from .pr_curve import PRCurve
 from .roc_curve import ROCCurve
@@ -8,5 +7,5 @@ from .average_precision_score import APscore, mAPscore
 from .roc_auc_score import AUCscore
 
 
-__all__ = ['BaseMetric', 'TopKAccuracy', 'ConfusionMatrix', 'CMAccuracy', 'CMPrecisionRecall',
+__all__ = ['TopKAccuracy', 'ConfusionMatrix', 'CMAccuracy', 'CMPrecisionRecall',
            'Fbetascore', 'PRCurve', 'ROCCurve', 'APscore', 'mAPscore', 'AUCscore']
