+++ volkscv/metrics/classification/fbeta_score.py
@@ -1,10 +1,8 @@
-import numpy as np
-
-from .base import BaseMetric
+from .base import BaseScoreCurve
 from .libs import fbeta_score
 
 
-class Fbetascore(BaseMetric):
+class Fbetascore(BaseScoreCurve):
     """
     Calculate Fbeta score for classification tasks, this method is restricted to
     the binary, multiclass, or multilabel-indicator classification task
@@ -71,10 +69,6 @@ class Fbetascore(BaseMetric):
         }
         super().__init__()
 
-    def reset(self):
-        self.y_true = None
-        self.probas_pred = None
-
     def compute(self, pred, target):
         self._y_true_temp = target
         if self.mode in ['binary', 'multiclass']:
@@ -85,17 +79,6 @@ class Fbetascore(BaseMetric):
             raise KeyError(f'mode "{self.mode}" do not exist')
 
         return None
-
-    def update(self, n=1):
-        def concatenator(total, temp):
-            if total is None:
-                total = temp
-            else:
-                total = np.concatenate((total, temp), axis=0)
-            return total
-
-        self.y_true = concatenator(self.y_true, self._y_true_temp)
-        self.probas_pred = concatenator(self.probas_pred, self._probas_pred_temp)
 
     def accumulate(self):
         accumulate_state = {}
@@ -118,7 +101,3 @@ class Fbetascore(BaseMetric):
             accumulate_state[f'f{self.beta}_score'] = f_score
 
         return accumulate_state
-
-    def check(self, pred, target):
-        super().check(pred, target)
-        self._check_pred_range(pred)
