+++ volkscv/metrics/classification/roc_auc_score.py
@@ -1,10 +1,8 @@
-import numpy as np
-
-from .base import BaseMetric
+from .base import BaseScoreCurve
 from .libs import roc_auc_score
 
 
-class AUCscore(BaseMetric):
+class AUCscore(BaseScoreCurve):
     """
     Calculate auc for classification tasks, this method is restricted to
     the binary, multiclass, or multilabel-indicator classification task
@@ -83,26 +81,6 @@ class AUCscore(BaseMetric):
         }
         super().__init__()
 
-    def reset(self):
-        self.y_true = None
-        self.probas_pred = None
-
-    def compute(self, pred, target):
-        self._y_true_temp = target
-        self._probas_pred_temp = pred
-        return None
-
-    def update(self, n=1):
-        def concatenator(total, temp):
-            if total is None:
-                total = temp
-            else:
-                total = np.concatenate((total, temp), axis=0)
-            return total
-
-        self.y_true = concatenator(self.y_true, self._y_true_temp)
-        self.probas_pred = concatenator(self.probas_pred, self._probas_pred_temp)
-
     def accumulate(self):
         accumulate_state = {}
 
@@ -127,6 +105,7 @@ class AUCscore(BaseMetric):
 
         return accumulate_state
 
-    def check(self, pred, target):
-        super().check(pred, target)
-        self._check_pred_range(pred)
+    @staticmethod
+    def _check_pred_sum(pred):
+        assert pred[0].sum() <= 1, \
+            "Pred should stand for the predicted probability with sum up to 1"
