+++ volkscv/metrics/segmentation/__init__.py
@@ -1,4 +1,3 @@
-from .base import BaseMetric
 from .confusion_matrix import ConfusionMatrix, Accuracy, IoU, mIoU, DiceScore
 
 __all__ = ['ConfusionMatrix', 'Accuracy', 'IoU', 'mIoU', 'DiceScore']
